pyproject.toml
[project]
name = "agent-framework-workshop"
version = "0.1.0"
description = "Proyecto para las demostraciones y laboratorios del workshop de Agent Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "agent-framework-core>=1.0.0b251001",
    "graphviz>=0.21",
    "ipykernel>=7.1.0",
    "mcp-server-filesystem>=0.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
